requirements.txt
colorama==0.4.6
decorator==5.2.1
greenlet==3.2.3
moviepy==2.2.1
numpy==2.3.0
playwright==1.52.0
proglog==0.1.12
pyee==13.0.0
python-decouple==3.8
python-dotenv==1.1.0
tenacity==9.1.2
typing_extensions==4.14.0